Change internal representation of register
Register is an object uniquely identified by its name (a character).
It has a few slots: name, value, print-func, jump-func and
insert-func that can be read by register-<SLOTNAME> selectors.

Function register-make, register-find and register-map along with the
selectors are public functions the rest of the code build on top of.

This has been submitted to http://tinyurl.com/4yth3cw. It may or may
not be included upstream.